+++ js/input-data.js
@@ -1,0 +1,38 @@
+const COUNT_PHOTOS = 25;
+const COUNT_AVATARS = 6;
+
+const MESSAGES = [
+  'Всё отлично!',
+  'В целом всё неплохо. Но не всё.',
+  'Когда вы делаете фотографию, хорошо бы убирать палец из кадра. В конце концов это просто непрофессионально.',
+  'Моя бабушка случайно чихнула с фотоаппаратом в руках и у неё получилась фотография лучше.',
+  'Я поскользнулся на банановой кожуре и уронил фотоаппарат на кота и у меня получилась фотография лучше.',
+  'Лица у людей на фотке перекошены, как будто их избивают. Как можно было поймать такой неудачный момент?!',
+];
+const NAMES = [
+  'Кристина',
+  'Макар',
+  'Лев',
+  'Анастасия',
+  'Леонид',
+  'Максим',
+  'Екатерина',
+  'Аделина',
+  'Павел',
+  'Ева',
+  'Егор',
+  'Мария',
+  'Игорь',
+  'Татьяна',
+  'Дмитрий',
+];
+const DESCRIPTION = [
+  'Обожаю море за то, что в нем водится рыба',
+  'Капсульный отель как раз мне по размеру',
+  'Моя Киса любит рыбку',
+  'Наш отдел выступает с докладом на всекотячей конференции',
+  'Общее фото отдела разработки. Кажется не хватает синих и зеленых цветов',
+  'Мой отпуск в Рыбатии',
+];
+
+export {COUNT_PHOTOS, COUNT_AVATARS, MESSAGES, NAMES, DESCRIPTION};
